Add `yardl init` and test Windows support (#8)
Adding a `yardl init PACKAGE_NAME` command that adds a _package.yml and
a sample model.yml file to a directory.

Additionally, adding a smoke test to our CI workflow that runs `yardl
init` and `yardl generate` and ensures we can build and run on Windows.
A few touch-ups were necessary to get the code building on Windows.

+++ tooling/internal/cpp/include/detail/hdf5/ddl.h
@@ -129,7 +129,7 @@ static inline H5::ArrayType FixedVectorDdl(H5::DataType const& element_type, hsi
  */
 static inline H5::ArrayType FixedNDArrayDdl(H5::DataType const& element_type,
                                             std::initializer_list<hsize_t> dimensions) {
-  return H5::ArrayType(element_type, dimensions.size(), std::data(dimensions));
+  return H5::ArrayType(element_type, static_cast<int>(dimensions.size()), std::data(dimensions));
 }
 
 /**
@@ -176,7 +176,7 @@ H5::EnumType UnionTypeEnumDdl(bool nullable, Labels const&... labels) {
 
 struct IndexEntry {
   int8_t type_;
-  u_int64_t offset_;
+  uint64_t offset_;
 };
 
 static inline H5::CompType UnionIndexDatasetElementTypeDdl(H5::EnumType type_enum) {
